requirements.txt
asgiref==3.7.2
certifi==2020.12.5
chardet==4.0.0
Django==4.2.2
django-cors-headers==4.1.0
djangorestframework==3.14.0
idna==2.10
numpy==1.25.0
psycopg2==2.9.6
pytz==2021.1
requests==2.25.1
shapely==2.0.1
sqlparse==0.4.4
typing_extensions==4.6.3
urllib3==1.26.4
